requirements.txt
Django==1.9.6
django-admin-sortable==2.0.15
django-allauth==0.25.2
django-cors-headers==1.1.0
django-modeltranslation==0.11
djangorestframework==3.3.3

git+https://github.com/vovaminof/django-rest-auth.git@roomit